refactored test app with sharing links
